refactor(anthropic): move #600 tests to their own file for the code-size ratchet
The new test module pushed lib.rs to 1231 LOC, over the 1200 LOC
oversized-file budget enforced by scripts/check_guardrails.sh. Split it
into trailing_assistant_repair_tests.rs, matching the pattern used
elsewhere (e.g. jcode-base browser_tests.rs).
